Entrega_5/Worker/requirements.txt
SQLAlchemy
marshmallow
psycopg2-binary
zipfile36
py7zr
bz2file
secure-smtplib
google-cloud-storage
flask
